Add simpler temporary error identification
