+++ go.mod
@@ -3,6 +3,7 @@ module github.com/brave-intl/challenge-bypass-server
 require (
 	github.com/actgardner/gogen-avro/v9 v9.0.0
 	github.com/aws/aws-sdk-go v1.27.0
+	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.13.0 // indirect
 	github.com/brave-intl/bat-go v0.4.0
 	github.com/brave-intl/challenge-bypass-ristretto-ffi v0.0.0-20190717223301-f88d942ddfaf
 	github.com/certifi/gocertifi v0.0.0-20180905225744-ee1a9a0726d2 // indirect
@@ -11,6 +12,7 @@ require (
 	github.com/go-chi/httplog v0.2.1 // indirect
 	github.com/golang-migrate/migrate/v4 v4.14.1
 	github.com/google/uuid v1.3.0 // indirect
+	github.com/jmespath/go-jmespath v0.4.0 // indirect
 	github.com/jmoiron/sqlx v1.2.0
 	github.com/lib/pq v1.9.0
 	github.com/patrickmn/go-cache v2.1.0+incompatible
